Crate slog_journald

Journald drain fro slog-rs

Since Journald supports structured data, structured data passed to slog is simply forwarded to Journald as structured data.

Examples

#[macro_use]
extern crate slog;
extern crate slog_journald;

use slog::*;
use slog_journald::*;

fn main() {
    let root = Logger::root(JournaldDrain.ignore_err(), o!("build_di" => "12344"));
    info!(root, "Testing journald"; "foo" => "bar");
}

Structs

JournaldDrain
Drain records and send to journald as structured data.

Enums

Error
Error type for logging to journald.